Added Enemy Manger Decoupled enemy spawning from Other things
Yeeee
